refactor(components): refactor space (#3757)
