+++ packages/components/space/index.ts
@@ -1,13 +1,9 @@
-import Space from './src/index'
+import { withInstall } from '@element-plus/utils/with-install'
 
-import type { App } from 'vue'
-import type { SFCWithInstall } from '@element-plus/utils/types'
+import Space from './src/space'
 
-const _Space = Space as SFCWithInstall<typeof Space>
+export const ElSpace = withInstall(Space)
+export default ElSpace
 
-_Space.install = (app: App) => {
-  app.component(_Space.name, _Space)
-}
-
-export default _Space
-export const ElSpace = _Space
+export * from './src/space'
+export * from './src/use-space'
