+++ packages/components/space/src/item.vue
@@ -1,22 +1,28 @@
 <template>
   <div :class="classes">
-    <slot></slot>
+    <slot />
   </div>
 </template>
 
 <script lang="ts">
 import { defineComponent, computed } from 'vue'
+import { buildProps } from '@element-plus/utils/props'
+
+const spaceItem = buildProps({
+  prefixCls: {
+    type: String,
+    default: 'el-space',
+  },
+} as const)
 
 export default defineComponent({
-  props: {
-    prefixCls: {
-      type: String,
-      default: 'el-space',
-    },
-  },
+  props: spaceItem,
+
   setup(props) {
+    const classes = computed(() => [`${props.prefixCls}__item`])
+
     return {
-      classes: computed(() => `${props.prefixCls}__item`),
+      classes,
     }
   },
 })
